libvmaf: stage libvmaf v2.0.0, soname v1.0.0
