+++ Makefile
@@ -1,7 +1,7 @@
 all:
 	cd third_party/libsvm && make lib
 
-	meson setup libvmaf/build libvmaf --buildtype release && \
+	meson setup libvmaf/build libvmaf --buildtype release -Denable_float=true && \
 	ninja -vC libvmaf/build
 	cd python && python3 setup.py build_ext --build-lib .
