+++ libvmaf/meson_options.txt
@@ -18,12 +18,12 @@ option('enable_avx512',
     value: false,
     description: 'Build AVX-512 asm files, requires nasm 2.14')
 
-option('install_rc',
-    type: 'boolean',
-    value: false,
-    description: 'Add rc libs/bins to list of installed files')
-
 option('built_in_models',
     type: 'boolean',
     value: true,
     description: 'Compile default vmaf models into the library')
+
+option('enable_float',
+    type: 'boolean',
+    value: false,
+    description: 'Compile floating-point feature extractors into the library')
